+++ HISTORY.rst
@@ -6,6 +6,7 @@ History
 v0.18 (unreleased)
 ..................
 * fix usage of ``max_burst_jobs``, improve coverage fix #152
+* stop lots of ``WatchVariableError`` errors in log, #153
 
 v0.17.1 (2019-08-21)
 ....................